requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
telethon>=1.30.0
pandas>=2.0.0
pyarrow>=12.0.0
streamlit>=1.30.0
gspread>=6.0.0
google-auth>=2.0.0